add feature to the Start Review button
